First model
Made it so that scraper copies embeddings from linkedin of posts that fit the category criteria.
Ollama_check_topic prompts Llama3.2 to check for context, unfortunately it's not that accurate.
Topics work best when given in a format "Topic1 , Topic2, Topic3,...", where model finds Topic1 most relevant, higher chance it will say the post is relevant if contains that topic.